・あたり判定に追加処理 ・アニメション関数作成 ・プレイヤなどのアニメ―ション関連の処理変更

+++ ActionRPG/cpp/Base.cpp
@@ -2,23 +2,20 @@
 
 //メンバイニシャライザにてメンバ変数初期化
 CharBase::CharBase() :
-	m_Pos{0.f},
-	m_Hp{0.f},
-	m_Speed{0.f}
-{
-}
+	m_Pos{ 0.f },
+	m_Hp{ 0 },
+	m_Speed{ 0.f }
+{}
 
-CharBase::CharBase(float x, float y, float z,float hp,float speed) :
+CharBase::CharBase(float x, float y, float z,int hp,float speed) :
 	m_Pos(VGet(x, y, z)),
 	m_Hp{ hp },
 	m_Speed{ speed }
-{
-}
+{}
 
-CharBase::CharBase(VECTOR pos, float hp, float speed) :
+CharBase::CharBase(VECTOR pos, int hp, float speed) :
 	m_Pos(pos),
 	m_Hp{ hp },
 	m_Speed{ speed }
-{
-}
+{}
